+++ src/datasource/DronerRankDatasource.ts
@@ -1,4 +1,8 @@
-import { DronerRankDetailEntity, DronerRankListEntity, taskDetailEntity } from './../entities/DronerRankEntities';
+import {
+  DronerRankDetailEntity,
+  DronerRankListEntity,
+  taskDetailEntity,
+} from "./../entities/DronerRankEntities";
 import { BASE_URL, httpClient } from "../config/config";
 export class DronerRankDatasource {
   static getDronerRank(
@@ -11,7 +15,9 @@ export class DronerRankDatasource {
     ratingMax?: number,
     startDate?: number,
     endDate?: number,
-    search?: string
+    search?: string,
+    sortDirection?: string,
+    sortField?: string
   ): Promise<DronerRankListEntity> {
     const params = {
       page: page,
@@ -24,6 +30,8 @@ export class DronerRankDatasource {
       startDate: startDate,
       endDate: endDate,
       search: search,
+      sortDirection: sortDirection,
+      sortField: sortField,
     };
     return httpClient
       .get(BASE_URL + "/tasks/droner-ranking/get-all-droner-ranking", {
@@ -36,9 +44,22 @@ export class DronerRankDatasource {
         console.log(error);
       });
   }
-  static getDronerRankById(id: string): Promise<DronerRankDetailEntity> {
+  static getDronerRankById(
+    id: string,
+    page: number,
+    row: number,
+    sortField?: string,
+    sortDirection?: string
+  ): Promise<DronerRankDetailEntity> {
+    const params = {
+      dronerId: id,
+      page: page,
+      take: row,
+      sortField:sortField,
+      sortDirection: sortDirection
+    }
     return httpClient
-      .get(BASE_URL + "/tasks/droner-ranking/get-droner-detail/" + id)
+      .get(BASE_URL + "/tasks/droner-ranking/get-droner-detail" ,{params})
       .then((response) => {
         return response.data;
       })